Finish settings base implementation
